+++ backend/EventTom/src/main/java/API/EventTom/DTO/TicketDTO.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
@@ -11,7 +12,7 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 public class TicketDTO {
-    public long finalPrice;
+    public BigDecimal finalPrice;
     public LocalDateTime purchaseDate;
     public boolean statusUsed;
     public long eventId;
