+++ backend/EventTom/src/main/java/API/EventTom/DTO/request/EventCreateDTO.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -13,9 +14,9 @@ public class EventCreateDTO {
     private String title;
     private String location;
     private LocalDateTime dateOfEvent;
-    private long totalTickets;
-    private long thresholdValue;
-    private long basePrice;
+    private int totalTickets;
+    private int thresholdValue;
+    private BigDecimal basePrice;
     private Long managerId;
 
 }
