+++ backend/EventTom/src/main/java/API/EventTom/DTO/request/EventUpdateDTO.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -12,7 +13,7 @@ import java.time.LocalDateTime;
 public class EventUpdateDTO {
     private String title;
     private LocalDateTime dateOfEvent;
-    private long totalTickets;
-    private long thresholdValue;
-    private long basePrice;
+    private int totalTickets;
+    private int thresholdValue;
+    private BigDecimal basePrice;
 }
